Split Registry interface into two classes.
`Registry` ABC now contains methods used by regular registry clients, this
interface is exposed via `Butler.registry` property. `ButlerRegistry` ABC
extends `Registry` with methods that are used by Butler, this includes
factory methods and a couple of internal methods. SQL and remote
implementations now inherit `ButlerRegistry`, they are not changed otherwise.

Also dropped `Registry.datasetIdFactory` attribute, we now have more transparent
method for generating dataset IDs and this attribute is not used.
