+++ python/lsst/daf/butler/registry/__init__.py
@@ -20,6 +20,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from . import interfaces, managers, queries, wildcards
+from ._butler_registry import ButlerRegistry
 from ._collection_summary import *
 from ._collectionType import *
 from ._config import *
