+++ python/lsst/daf/butler/tests/_datasetsHelper.py
@@ -37,7 +37,8 @@ from lsst.daf.butler import DataCoordinate, DatasetRef, DatasetType, StorageClas
 from lsst.daf.butler.formatters.yaml import YamlFormatter
 
 if TYPE_CHECKING:
-    from lsst.daf.butler import Config, DatasetId, Datastore, Dimension, DimensionGraph, Registry
+    from lsst.daf.butler import Config, DatasetId, Datastore, Dimension, DimensionGraph
+    from lsst.daf.butler.registry import ButlerRegistry
 
 
 class DatasetTestHelper:
@@ -101,7 +102,7 @@ class DatastoreTestHelper:
     datastoreType: type[Datastore]
     configFile: str
 
-    def setUpDatastoreTests(self, registryClass: type[Registry], configClass: type[Config]) -> None:
+    def setUpDatastoreTests(self, registryClass: type[ButlerRegistry], configClass: type[Config]) -> None:
         """Shared setUp code for all Datastore tests."""
         self.registry = registryClass()
         self.config = configClass(self.configFile)
