+++ tests/test_dimensions.py
@@ -42,11 +42,10 @@ from lsst.daf.butler import (
     DimensionUniverse,
     NamedKeyDict,
     NamedValueSet,
-    Registry,
     TimespanDatabaseRepresentation,
     YamlRepoImportBackend,
 )
-from lsst.daf.butler.registry import RegistryConfig
+from lsst.daf.butler.registry import ButlerRegistry, RegistryConfig
 
 DIMENSION_DATA_FILE = os.path.normpath(
     os.path.join(os.path.dirname(__file__), "data", "registry", "hsc-rc2-subset.yaml")
@@ -65,7 +64,7 @@ def loadDimensionData() -> DataCoordinateSequence:
     # data and retreive it as a set of DataCoordinate objects.
     config = RegistryConfig()
     config["db"] = "sqlite://"
-    registry = Registry.createFromConfig(config)
+    registry = ButlerRegistry.createFromConfig(config)
     with open(DIMENSION_DATA_FILE) as stream:
         backend = YamlRepoImportBackend(stream, registry)
     backend.register()
