+++ tests/test_quantumBackedButler.py
@@ -34,10 +34,10 @@ from lsst.daf.butler import (
     Quantum,
     QuantumBackedButler,
     QuantumProvenanceData,
-    Registry,
     RegistryConfig,
     StorageClass,
 )
+from lsst.daf.butler.registry import ButlerRegistry
 from lsst.daf.butler.tests.utils import makeTestTempDir, removeTestTempDir
 from lsst.resources import ResourcePath
 
@@ -55,7 +55,7 @@ class QuantumBackedButlerTestCase(unittest.TestCase):
 
         # Make a butler and import dimension definitions.
         registryConfig = RegistryConfig(self.config.get("registry"))
-        Registry.createFromConfig(registryConfig, butlerRoot=self.root)
+        ButlerRegistry.createFromConfig(registryConfig, butlerRoot=self.root)
         self.butler = Butler(self.config, writeable=True, run="RUN")
         self.butler.import_(filename=os.path.join(TESTDIR, "data", "registry", "base.yaml"))
